UNTESTED CHANGES - My local test server is not relecting update to source files, even when restarted. Please test before merging.
Changed the Physician fields to be completely hidden when No is selected for the Was Physician Notified field.
Updated the implementation of the Show Hidden Fields button so that it should be hidden when sumbitting a new error, but not when loading any error. Test for the button was changed to "Show currently enabled fields" to better reflect the behaviour.
Made the Was Physician Notified selection actually dependant on the new ShowPhysicianNotified toggle. Added a line to set the field to the value of ShowPhysicianFields when loading a blank form or clicking the Show Currently Enabled Fields button.